Rename app:environment:* to env:*.
