+++ src/Command/Environment/EnvironmentMirrorCommand.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Acquia\Cli\Command\App;
+namespace Acquia\Cli\Command\Environment;
 
 use Acquia\Cli\Command\CommandBase;
 use Acquia\Cli\Output\Checklist;
@@ -18,7 +18,7 @@ use Symfony\Component\Console\Output\OutputInterface;
  */
 class EnvironmentMirrorCommand extends CommandBase {
 
-  protected static $defaultName = 'app:environment:mirror';
+  protected static $defaultName = 'env:mirror';
 
   /**
    * @var \Acquia\Cli\Output\Checklist
