+++ tests/phpunit/src/Commands/Environment/CopyCronTasksCommandTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Acquia\Cli\Tests\Commands\App;
+namespace Acquia\Cli\Tests\Commands\Environment;
 
 use Acquia\Cli\Command\App\CopyCronTasksCommand;
 use Acquia\Cli\Tests\CommandTestBase;
